requirements.txt
aiogram==3.4.0
sqlalchemy==1.4.22
psycopg2-binary==2.9.9
python-dotenv==0.19.0
google-generativeai==0.4.1
loguru==0.7.2
asyncpg==0.29.0
asyncpg==0.29.0